Improve 'Testing with your app' contributing docs
While trying to run spring locally to test a bug in the Basecamp app I
found the instructions confusing and missing important details that you
normally don't need to deal with when working with local gems.

It was confusing because when I would run `spring stop` I'd see I was
linked to my local copy (with a puts in initialize) but then `spring
help` was still linked to spring 1.3.3 and not my local copy.

Documentation was updated to include installing the gem in the application
you are testing against via the path and ensuring you update the
`Gemfile.lock` first if your local copy version number does not match the
`Gemfile.lock` version.